src/requirements.txt
# install with 
# conda install --file requirements.txt
pytorch::pytorch
torchvision
torchaudio
numpy
matplotlib
tensorboard
tqdm